added basic components to stats page
