+++ src/app/s/dash/stats.tsx
@@ -1,0 +1,27 @@
+"use client"
+
+import Typography from "@/_components/Typography";
+import Card from "@/_components/Card";
+
+export default function Stats() {
+
+  return (
+    <>
+      <Card className="w-3/5 inline-block" title="Your Karma Score: 80.4">
+        ee
+      </Card>
+      <div className="w-2/5 inline-block">
+        <div className="bg-primary m-4 rounded-2xl flex flex-col justify-center p-4 ">
+          <div className="flex items-center h-full mb-3">
+            <Typography variant="h1">
+              $273.42
+              feferferfe
+              rferfe
+            </Typography>
+          </div>
+        </div>
+      </div>
+
+    </>
+  )
+}
